final architecture
