+++ custom_bikes/funciones/funciones_clientes.py
@@ -1,0 +1,3 @@
+def insert_cliente():
+    print('clientes')
+
