+++ custom_bikes/menu/menu_stock.py
