--- funciones/funciones_clientes.py
@@ -1,2 +1,0 @@
-def agregar_venta_manual():
-
